also pull out centerTruncate too
