+++ static/app/views/insights/common/components/releaseSelector.tsx
@@ -26,7 +26,7 @@ import {
   useReleases,
   useReleaseSelection,
 } from 'sentry/views/insights/common/queries/useReleases';
-import {formatVersionAndCenterTruncate} from 'sentry/views/insights/common/utils/centerTruncate';
+import {formatVersionAndCenterTruncate} from 'sentry/views/insights/common/utils/formatVersionAndCenterTruncate';
 
 export const PRIMARY_RELEASE_ALIAS = 'R1';
 export const SECONDARY_RELEASE_ALIAS = 'R2';
